+++ tests/test_sparse_command.rs
@@ -12,15 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::io::Write;
+
 use crate::common::TestEnvironment;
 
 pub mod common;
 
 #[test]
 fn test_sparse_manage_patterns() {
-    let test_env = TestEnvironment::default();
+    let mut test_env = TestEnvironment::default();
     test_env.jj_cmd_success(test_env.env_root(), &["init", "repo", "--git"]);
     let repo_path = test_env.env_root().join("repo");
+
+    let edit_script = test_env.set_up_fake_editor();
 
     // Write some files to the working copy
     std::fs::write(repo_path.join("file1"), "contents").unwrap();
@@ -114,4 +118,54 @@ fn test_sparse_manage_patterns() {
     assert!(repo_path.join("file1").exists());
     assert!(repo_path.join("file2").exists());
     assert!(repo_path.join("file3").exists());
+
+    // Can edit with editor
+    let edit_patterns = |patterns: &[&str]| {
+        let mut file = std::fs::File::create(&edit_script).unwrap();
+        file.write_all(b"dump patterns0\0write\n").unwrap();
+        for pattern in patterns {
+            file.write_all(pattern.as_bytes()).unwrap();
+            file.write_all(b"\n").unwrap();
+        }
+    };
+    let read_patterns = || std::fs::read_to_string(test_env.env_root().join("patterns0")).unwrap();
+
+    edit_patterns(&["file1"]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["sparse", "set", "--edit"]);
+    insta::assert_snapshot!(stdout, @r###"
+    Added 0 files, modified 0 files, removed 2 files
+    "###);
+    insta::assert_snapshot!(read_patterns(), @".");
+    let stdout = test_env.jj_cmd_success(&repo_path, &["sparse", "list"]);
+    insta::assert_snapshot!(stdout, @r###"
+    file1
+    "###);
+
+    // Can edit with `--clear` and `--add`
+    edit_patterns(&["file2"]);
+    let stdout = test_env.jj_cmd_success(
+        &repo_path,
+        &["sparse", "set", "--edit", "--clear", "--add", "file1"],
+    );
+    insta::assert_snapshot!(stdout, @r###"
+    Added 1 files, modified 0 files, removed 1 files
+    "###);
+    insta::assert_snapshot!(read_patterns(), @"file1");
+    let stdout = test_env.jj_cmd_success(&repo_path, &["sparse", "list"]);
+    insta::assert_snapshot!(stdout, @r###"
+    file2
+    "###);
+
+    // Can edit with multiple files
+    edit_patterns(&["file2", "file3"]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["sparse", "set", "--clear", "--edit"]);
+    insta::assert_snapshot!(stdout, @r###"
+    Added 1 files, modified 0 files, removed 0 files
+    "###);
+    insta::assert_snapshot!(read_patterns(), @"");
+    let stdout = test_env.jj_cmd_success(&repo_path, &["sparse", "list"]);
+    insta::assert_snapshot!(stdout, @r###"
+    file2
+    file3
+    "###);
 }